Initial commit
This is the initial commit of a kata-webhook fork.  The reason for
forking is nothing else than switching to using Go Modules, as what we
have as part of the current `kata-containers/tests` repository limits us
to no update to newer versions of the kubewebhook, which may (or may
not) contain fixes that are interesting for us.

This is kata-webhook as of
https://github.com/kata-containers/tests/commit/3018a09e694520fde3de72042a3f7b91cfe5427c

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>